Phase1/requirements.txt
google-cloud-pubsub==2.4.1
jsonschema==3.2.0
pyyaml==5.4.1
argparse==1.4.0